in process of refactoring state storage
